refactor: signatures
